+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Version 16
+
+* Added crtp-pattern helper
+
 # Version 15
 
 * Updated signal interface